+++ test/assets/CustomModuleName/CustomModuleName.csproj
@@ -4,7 +4,6 @@
     <TargetFrameworks>net8.0;net9.0;net10.0</TargetFrameworks>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
-    <RootNamespace>CustomRegistrationMethodName</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
